replace txqueue
