+++ block-producer/src/tasks/mod.rs
@@ -1,3 +1,1 @@
 pub mod txqueue;
-
-pub mod txqueuerpc;
