+++ block-producer/src/tasks/txqueue/tests/batch_sent.rs
@@ -6,39 +6,29 @@ use super::*;
 /// Transaction timer is set to infinite. After 50ms delay, we confirm that the batch was sent
 #[tokio::test]
 async fn test_batch_full_sent() {
-    let interval_duration: Duration = Duration::from_millis(10);
-    let batch_size: usize = 3;
+    let batch_size = 3;
 
-    let handle_in = {
-        let num_txs_to_send = batch_size;
-
-        HandleInFixedInterval::new(interval_duration, num_txs_to_send)
-    };
-    let handle_out = HandleOutDefault::new();
-
-    let tx_queue = TxQueue::new(
-        handle_in,
-        handle_out,
+    let (read_tx_client, ready_queue, batches) = setup(
+        VerifyTxRpcSuccess,
         TxQueueOptions {
             batch_size,
             tx_max_time_in_queue: Duration::MAX,
         },
     );
 
-    // Spawn tx_queue task
-    {
-        let tx_queue = tx_queue.clone();
-        tokio::spawn(tx_queue.run());
-    }
+    // Start fixed interval client
+    tokio::spawn(
+        ReadTxClientFixedInterval::new(read_tx_client, Duration::from_millis(10), batch_size).run(),
+    );
 
     time::sleep(Duration::from_millis(50)).await;
 
     // Ensure that the batch was sent
-    assert_eq!(tx_queue.handle_out.batches.lock().await.len(), 1);
+    assert_eq!(batches.lock().await.len(), 1);
     // Ensure that the batch contains all the transactions
-    assert_eq!(tx_queue.handle_out.batches.lock().await[0].len(), batch_size);
+    assert_eq!(batches.lock().await[0].len(), batch_size);
     // Ensure that the queue is empty
-    assert!(tx_queue.ready_queue.lock().await.is_empty());
+    assert!(ready_queue.lock().await.is_empty());
 }
 
 /// Tests that if `num_txs_in_queue > batch_size`, we still send a batch of
@@ -48,14 +38,10 @@ async fn test_batch_full_sent() {
 /// immediately check if a batch is ready.
 #[tokio::test]
 async fn test_proper_batch_size_sent() {
-    let batch_size: usize = 3;
+    let batch_size = 3;
 
-    let handle_in = HandleInVariableInterval::new(vec![Duration::from_millis(30)]);
-    let handle_out = HandleOutDefault::new();
-
-    let tx_queue = TxQueue::new(
-        handle_in,
-        handle_out,
+    let (read_tx_client, ready_queue, batches) = setup(
+        VerifyTxRpcSuccess,
         TxQueueOptions {
             batch_size,
             tx_max_time_in_queue: Duration::MAX,
@@ -65,27 +51,26 @@ async fn test_proper_batch_size_sent() {
     // Fill queue so that `queue_size == batch_size`
     {
         let proven_tx_generator = DummyProvenTxGenerator::new();
-        let mut locked_queue = tx_queue.ready_queue.lock().await;
+        let mut locked_queue = ready_queue.lock().await;
 
         locked_queue.push(Arc::new(proven_tx_generator.dummy_proven_tx()));
         locked_queue.push(Arc::new(proven_tx_generator.dummy_proven_tx()));
         locked_queue.push(Arc::new(proven_tx_generator.dummy_proven_tx()));
     }
 
-    // Spawn tx_queue task
-    {
-        let tx_queue = tx_queue.clone();
-        tokio::spawn(tx_queue.run());
-    }
+    // Start client that sends 1 transaction after 10ms
+    tokio::spawn(
+        ReadTxClientVariableInterval::new(read_tx_client, vec![Duration::from_millis(10)]).run(),
+    );
 
     time::sleep(Duration::from_millis(50)).await;
 
     // Ensure that the batch was sent
-    assert_eq!(tx_queue.handle_out.batches.lock().await.len(), 1);
+    assert_eq!(batches.lock().await.len(), 1);
     // Ensure that the batch contains `batch_size` elements
-    assert_eq!(tx_queue.handle_out.batches.lock().await[0].len(), batch_size);
+    assert_eq!(batches.lock().await[0].len(), batch_size);
     // Ensure that the queue contains 1 transaction
-    assert_eq!(tx_queue.ready_queue.lock().await.len(), 1);
+    assert_eq!(ready_queue.lock().await.len(), 1);
 }
 
 /// Tests that when a transaction's verification fails, it is not added to the queue.
@@ -94,37 +79,27 @@ async fn test_proper_batch_size_sent() {
 /// Transaction timer is set to 10ms. After 50ms delay, we confirm that no batch was sent.
 #[tokio::test]
 async fn test_tx_verification_failure() {
-    let interval_duration: Duration = Duration::from_millis(10);
-    let batch_size: usize = 3;
+    let batch_size = 3;
 
-    let handle_in = {
-        let num_txs_to_send = batch_size;
-
-        HandleInFixedInterval::new(interval_duration, num_txs_to_send)
-    };
-    let handle_out = HandleOutFailVerification::new();
-
-    let tx_queue = TxQueue::new(
-        handle_in,
-        handle_out,
+    let (read_tx_client, ready_queue, batches) = setup(
+        VerifyTxRpcFailure,
         TxQueueOptions {
             batch_size,
             tx_max_time_in_queue: Duration::MAX,
         },
     );
 
-    // Spawn tx_queue task
-    {
-        let tx_queue = tx_queue.clone();
-        tokio::spawn(tx_queue.run());
-    }
+    // Start fixed interval client
+    tokio::spawn(
+        ReadTxClientFixedInterval::new(read_tx_client, Duration::from_millis(10), batch_size).run(),
+    );
 
     time::sleep(Duration::from_millis(50)).await;
 
     // Ensure that no batch was sent
-    assert!(tx_queue.handle_out.batches.lock().await.is_empty());
+    assert!(batches.lock().await.is_empty());
     // Ensure that the queue is empty
-    assert!(tx_queue.ready_queue.lock().await.is_empty());
+    assert!(ready_queue.lock().await.is_empty());
 }
 
 /// Tests that if a batch is not full, then the batch will be sent regardless
@@ -137,35 +112,33 @@ async fn test_tx_verification_failure() {
 /// first transaction (as opposed to possibly the second one).
 #[tokio::test]
 async fn test_timer_send_batch() {
-    let batch_size: usize = 3;
+    let batch_size = 3;
 
-    let handle_in =
-        HandleInVariableInterval::new(vec![Duration::from_millis(0), Duration::from_millis(20)]);
-    let handle_out = HandleOutDefault::new();
-
-    let tx_queue = TxQueue::new(
-        handle_in,
-        handle_out,
+    let (read_tx_client, ready_queue, batches) = setup(
+        VerifyTxRpcSuccess,
         TxQueueOptions {
             batch_size,
             tx_max_time_in_queue: Duration::from_millis(30),
         },
     );
 
-    // Spawn tx_queue task
-    {
-        let tx_queue = tx_queue.clone();
-        tokio::spawn(tx_queue.run());
-    }
+    // Start client
+    tokio::spawn(
+        ReadTxClientVariableInterval::new(
+            read_tx_client,
+            vec![Duration::from_millis(0), Duration::from_millis(20)],
+        )
+        .run(),
+    );
 
     time::sleep(Duration::from_millis(40)).await;
 
     // Ensure that the batch was sent
-    assert_eq!(tx_queue.handle_out.batches.lock().await.len(), 1);
+    assert_eq!(batches.lock().await.len(), 1);
     // Ensure that the batch contains all the transactions
-    assert_eq!(tx_queue.handle_out.batches.lock().await[0].len(), batch_size - 1);
+    assert_eq!(batches.lock().await[0].len(), batch_size - 1);
     // Ensure that the queue is empty
-    assert!(tx_queue.ready_queue.lock().await.is_empty());
+    assert!(ready_queue.lock().await.is_empty());
 }
 
 /// This tests if the internal transaction timer properly resets after a full
@@ -190,45 +163,40 @@ async fn test_timer_send_batch() {
 async fn test_tx_timer_resets_after_full_batch_sent() {
     let batch_size: usize = 2;
 
-    let handle_in = HandleInVariableInterval::new(vec![
-        Duration::from_millis(0),
-        Duration::from_millis(10),
-        Duration::from_millis(20),
-    ]);
-    let handle_out = HandleOutDefault::new();
-
-    let tx_queue = TxQueue::new(
-        handle_in,
-        handle_out,
+    let (read_tx_client, ready_queue, batches) = setup(
+        VerifyTxRpcSuccess,
         TxQueueOptions {
             batch_size,
             tx_max_time_in_queue: Duration::from_millis(40),
         },
     );
 
-    // Spawn tx_queue task
-    {
-        let tx_queue = tx_queue.clone();
-        tokio::spawn(tx_queue.run());
-    }
+    // Start client
+    tokio::spawn(
+        ReadTxClientVariableInterval::new(
+            read_tx_client,
+            vec![Duration::from_millis(0), Duration::from_millis(10), Duration::from_millis(20)],
+        )
+        .run(),
+    );
 
     time::sleep(Duration::from_millis(55)).await;
 
     // Ensure that 1 batch was sent
-    assert_eq!(tx_queue.handle_out.batches.lock().await.len(), 1);
+    assert_eq!(batches.lock().await.len(), 1);
     // Ensure that the queue holds 1 transaction
-    assert_eq!(tx_queue.ready_queue.lock().await.len(), 1);
+    assert_eq!(ready_queue.lock().await.len(), 1);
 
     time::sleep(Duration::from_millis(35)).await;
 
     // Ensure that 2 batches were sent
-    assert_eq!(tx_queue.handle_out.batches.lock().await.len(), 2);
+    assert_eq!(batches.lock().await.len(), 2);
     // Ensure that first batch received is length 2
-    assert_eq!(tx_queue.handle_out.batches.lock().await[0].len(), 2);
+    assert_eq!(batches.lock().await[0].len(), 2);
     // Ensure that second batch received is length 1
-    assert_eq!(tx_queue.handle_out.batches.lock().await[1].len(), 1);
+    assert_eq!(batches.lock().await[1].len(), 1);
     // Ensure that the queue is empty
-    assert!(tx_queue.ready_queue.lock().await.is_empty());
+    assert!(ready_queue.lock().await.is_empty());
 }
 
 /// This tests if the internal transaction timer properly resets after a
@@ -254,43 +222,72 @@ async fn test_tx_timer_resets_after_full_batch_sent() {
 /// wouldn't be the case if we forgot to reset the timer after tx1 is sent in a
 /// batch.
 #[tokio::test]
-async fn test_tx_timer_resets_timeout() {
+async fn test_tx_timer_resets_after_timeout() {
     let batch_size: usize = 2;
 
-    let handle_in =
-        HandleInVariableInterval::new(vec![Duration::from_millis(0), Duration::from_millis(45)]);
-    let handle_out = HandleOutDefault::new();
-
-    let tx_queue = TxQueue::new(
-        handle_in,
-        handle_out,
+    let (read_tx_client, ready_queue, batches) = setup(
+        VerifyTxRpcSuccess,
         TxQueueOptions {
             batch_size,
             tx_max_time_in_queue: Duration::from_millis(30),
         },
     );
 
-    // Spawn tx_queue task
-    {
-        let tx_queue = tx_queue.clone();
-        tokio::spawn(tx_queue.run());
-    }
+    // Start client
+    tokio::spawn(
+        ReadTxClientVariableInterval::new(
+            read_tx_client,
+            vec![Duration::from_millis(0), Duration::from_millis(45)],
+        )
+        .run(),
+    );
 
     time::sleep(Duration::from_millis(75)).await;
 
     // Ensure that 1 batch was sent
-    assert_eq!(tx_queue.handle_out.batches.lock().await.len(), 1);
+    assert_eq!(batches.lock().await.len(), 1);
     // Ensure that the queue holds 1 transaction
-    assert_eq!(tx_queue.ready_queue.lock().await.len(), 1);
+    assert_eq!(ready_queue.lock().await.len(), 1);
 
     time::sleep(Duration::from_millis(35)).await;
 
     // Ensure that 2 batches were sent
-    assert_eq!(tx_queue.handle_out.batches.lock().await.len(), 2);
+    assert_eq!(batches.lock().await.len(), 2);
     // Ensure that first batch received is length 1
-    assert_eq!(tx_queue.handle_out.batches.lock().await[0].len(), 1);
+    assert_eq!(batches.lock().await[0].len(), 1);
     // Ensure that second batch received is length 1
-    assert_eq!(tx_queue.handle_out.batches.lock().await[1].len(), 1);
+    assert_eq!(batches.lock().await[1].len(), 1);
     // Ensure that the queue is empty
-    assert!(tx_queue.ready_queue.lock().await.is_empty());
+    assert!(ready_queue.lock().await.is_empty());
+}
+
+// HELPERS
+// ================================================================================================
+
+/// Starts the RPC servers (txqueue's server and servers which txqueue is a client).
+/// Returns handles useful for tests
+fn setup<VerifyTxServerImpl>(
+    verify_tx_server_impl: VerifyTxServerImpl,
+    tx_queue_options: TxQueueOptions,
+) -> (ReadTxRpcClient, ReadyQueue, SharedMutVec<Vec<SharedProvenTx>>)
+where
+    VerifyTxServerImpl: ServerImpl<SharedProvenTx, Result<(), VerifyTxError>>,
+{
+    let (verify_tx_client, verify_tx_server) = create_client_server_pair(verify_tx_server_impl);
+
+    let send_txs_server_impl = SendTxsDefaultServerImpl::new();
+    let batches = send_txs_server_impl.batches.clone();
+    let (send_txs_client, send_txs_server) = create_client_server_pair(send_txs_server_impl);
+
+    let tx_queue = TxQueue::new(verify_tx_client, send_txs_client, tx_queue_options);
+
+    let ready_queue = tx_queue.ready_queue.clone();
+    let (read_tx_client, read_tx_server) = create_client_server_pair(tx_queue);
+
+    // Start servers
+    tokio::spawn(verify_tx_server.serve());
+    tokio::spawn(send_txs_server.serve());
+    tokio::spawn(read_tx_server.serve());
+
+    (read_tx_client, ready_queue, batches)
 }
